requirements.txt
pandas
plotly
kaleido
pillow